moved miri scattering into detector added charge difussion values based on comparisons other small fixes
